not actually computer science, but still computing
cf. Birnbaum and Langmead (2017): "The first step toward learning to code is to recognize that computer programming is not computer science; it is more like writing. Everyone can learn to do it, and can be given the opportunity to learn to do it in ways that are appropriate for their disciplines" (76).

https://doi.org/10.1007/978-3-319-54226-3_5